pyproject.toml
[tool.poetry]
name = "envsarujuk"
version = "0.1.0"
description = ""
authors = ["Michkail <55208634+Michkail@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
whitenoise = "^6.6.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
